Refactor CentralLimitTheorm graph
This commit breaks up CentralLimitGraph into its components.
