+++ src/central_limit_theorem/SampleRangeSlider.jsx
@@ -1,0 +1,36 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { forceNumber } from '../utils';
+
+export const SampleRangeSlider = ({enableSampleSlider, numberOfSamples,
+    sampleMeansIdx, handleSampleMeansIdx, runSample}) => {
+    const handleSampleMeans = (e) => {
+        e.preventDefault();
+        handleSampleMeansIdx(forceNumber(e.target.value));
+    };
+    return (
+        <>
+        <form>
+            <fieldset>
+                <legend>Step 3: Observe the changes among samples</legend>
+                <div>
+                    <input type="range"
+                        id="sample-slider"
+                        disabled={ enableSampleSlider ? false : true}
+                        min="1"
+                        max={numberOfSamples}
+                        value={sampleMeansIdx}
+                        onChange={handleSampleMeans} />
+                </div>
+            </fieldset>
+        </form>
+        </>
+    );
+};
+
+SampleRangeSlider.propTypes = {
+    enableSampleSlider: PropTypes.bool,
+    numberOfSamples: PropTypes.number,
+    sampleMeansIdx: PropTypes.number,
+    handleSampleMeansIdx: PropTypes.func,
+};
